Cluster-autoscaler: move ToBeDeleted taint functions to utils
